refactor(server): split fixed-size converter file
