+++ server/src/core/converter/escape-csv.ts
@@ -1,0 +1,5 @@
+export function escapeCsv(value: string): string {
+  if (/[",\n\r]/.test(value)) return '"' + value.replaceAll('"', '""') + '"';
+
+  return value;
+}
